# apps/ts/packages/shared/openapi/bt-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "trading25-bt API",
    "description": "バックテスト実行のためのREST API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/strategies/{strategy_name}": {
      "get": {
        "tags": [
          "Strategies"
        ],
        "summary": "Get Strategy Detail",
        "description": "戦略詳細を取得\n\nArgs:\n    strategy_name: 戦略名（例: 'range_break_v5', 'production/range_break_v5'）",
        "operationId": "get_strategy_detail_api_strategies__strategy_name__get",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Strategies"
        ],
        "summary": "Update Strategy",
        "description": "戦略設定を更新\n\nArgs:\n    strategy_name: 戦略名\n    request: 更新する設定\n\nNote:\n    experimentalカテゴリのみ更新可能",
        "operationId": "update_strategy_api_strategies__strategy_name__put",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Strategies"
        ],
        "summary": "Delete Strategy",
        "description": "戦略を削除\n\nArgs:\n    strategy_name: 戦略名\n\nNote:\n    experimentalカテゴリのみ削除可能",
        "operationId": "delete_strategy_api_strategies__strategy_name__delete",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDeleteResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_name}/validate": {
      "post": {
        "tags": [
          "Strategies"
        ],
        "summary": "Validate Strategy",
        "description": "戦略設定を検証\n\nArgs:\n    strategy_name: 戦略名\n    request: 検証する設定（省略時は既存の設定を検証）",
        "operationId": "validate_strategy_api_strategies__strategy_name__validate_post",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/StrategyValidationRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Request"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyValidationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_name}/duplicate": {
      "post": {
        "tags": [
          "Strategies"
        ],
        "summary": "Duplicate Strategy",
        "description": "戦略を複製\n\nArgs:\n    strategy_name: 複製元の戦略名\n    request: 複製先の新しい戦略名\n\nNote:\n    複製先は常にexperimentalカテゴリ",
        "operationId": "duplicate_strategy_api_strategies__strategy_name__duplicate_post",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyDuplicateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyDuplicateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/strategies/{strategy_name}/rename": {
      "post": {
        "tags": [
          "Strategies"
        ],
        "summary": "Rename Strategy",
        "description": "戦略をリネーム\n\nArgs:\n    strategy_name: 現在の戦略名\n    request: リネームリクエスト（新しい戦略名）\n\nNote:\n    experimentalカテゴリのみリネーム可能",
        "operationId": "rename_strategy_api_strategies__strategy_name__rename_post",
        "parameters": [
          {
            "name": "strategy_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Strategy Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyRenameResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/config/default": {
      "get": {
        "tags": [
          "Strategies"
        ],
        "summary": "Get Default Config",
        "description": "デフォルト設定をraw YAML文字列として取得\n\nコメントを保持するためYAML文字列をそのまま返却",
        "operationId": "get_default_config_api_config_default_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultConfigResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Strategies"
        ],
        "summary": "Update Default Config",
        "description": "デフォルト設定を更新\n\nYAML文字列を受け取り、パース検証後にdefault.yamlに書き込み、\nConfigLoaderのメモリ内デフォルト設定をリロード",
        "operationId": "update_default_config_api_config_default_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DefaultConfigUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DefaultConfigUpdateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DefaultConfigResponse": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "default.yamlの内容（YAML文字列）"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "DefaultConfigResponse",
        "description": "デフォルト設定レスポンス（raw YAML文字列）"
      },
      "DefaultConfigUpdateRequest": {
        "properties": {
          "content": {
            "type": "string",
            "title": "Content",
            "description": "更新するYAML文字列"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "DefaultConfigUpdateRequest",
        "description": "デフォルト設定更新リクエスト"
      },
      "DefaultConfigUpdateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "更新成功フラグ"
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "DefaultConfigUpdateResponse",
        "description": "デフォルト設定更新レスポンス"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyDeleteResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "削除成功フラグ"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "削除した戦略名"
          }
        },
        "type": "object",
        "required": [
          "success",
          "strategy_name"
        ],
        "title": "StrategyDeleteResponse",
        "description": "戦略削除レスポンス"
      },
      "StrategyDetailResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "description": "戦略名"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "description": "カテゴリ"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "表示名"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "説明"
          },
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "戦略設定（YAML）"
          },
          "execution_info": {
            "additionalProperties": true,
            "type": "object",
            "title": "Execution Info",
            "description": "実行情報"
          }
        },
        "type": "object",
        "required": [
          "name",
          "category",
          "config",
          "execution_info"
        ],
        "title": "StrategyDetailResponse",
        "description": "戦略詳細レスポンス"
      },
      "StrategyDuplicateRequest": {
        "properties": {
          "new_name": {
            "type": "string",
            "title": "New Name",
            "description": "新しい戦略名（カテゴリなし）"
          }
        },
        "type": "object",
        "required": [
          "new_name"
        ],
        "title": "StrategyDuplicateRequest",
        "description": "戦略複製リクエスト"
      },
      "StrategyDuplicateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "複製成功フラグ"
          },
          "new_strategy_name": {
            "type": "string",
            "title": "New Strategy Name",
            "description": "新しい戦略名"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "保存先パス"
          }
        },
        "type": "object",
        "required": [
          "success",
          "new_strategy_name",
          "path"
        ],
        "title": "StrategyDuplicateResponse",
        "description": "戦略複製レスポンス"
      },
...
      "StrategyRenameRequest": {
        "properties": {
          "new_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "New Name",
            "description": "新しい戦略名（カテゴリなし）"
          }
        },
        "type": "object",
        "required": [
          "new_name"
        ],
        "title": "StrategyRenameRequest",
        "description": "戦略リネームリクエスト"
      },
      "StrategyRenameResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "リネーム成功フラグ"
          },
          "old_name": {
            "type": "string",
            "title": "Old Name",
            "description": "変更前の戦略名"
          },
          "new_name": {
            "type": "string",
            "title": "New Name",
            "description": "変更後の戦略名"
          },
          "new_path": {
            "type": "string",
            "title": "New Path",
            "description": "新しいファイルパス"
          }
        },
        "type": "object",
        "required": [
          "success",
          "old_name",
          "new_name",
          "new_path"
        ],
        "title": "StrategyRenameResponse",
        "description": "戦略リネームレスポンス"
      },
      "StrategyUpdateRequest": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "更新する戦略設定"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "StrategyUpdateRequest",
        "description": "戦略設定更新リクエスト"
      },
      "StrategyUpdateResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "description": "更新成功フラグ"
          },
          "strategy_name": {
            "type": "string",
            "title": "Strategy Name",
            "description": "戦略名"
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "保存先パス"
          }
        },
        "type": "object",
        "required": [
          "success",
          "strategy_name",
          "path"
        ],
        "title": "StrategyUpdateResponse",
        "description": "戦略設定更新レスポンス"
      },
      "StrategyValidationRequest": {
        "properties": {
          "config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Config",
            "description": "検証する戦略設定"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "StrategyValidationRequest",
        "description": "戦略設定検証リクエスト"
      },
      "StrategyValidationResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "description": "検証結果"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "エラーメッセージ一覧"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings",
            "description": "警告メッセージ一覧"
          }
        },
        "type": "object",
        "required": [
          "valid"
        ],
        "title": "StrategyValidationResponse",
        "description": "戦略設定検証レスポンス"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}